Move function from trigger/utils into module using it
